+++ app/src/main/java/au/com/shiftyjelly/pocketcasts/ui/MainActivity.kt
@@ -687,7 +687,13 @@ class MainActivity :
                 updateNavAndStatusColors(true, state.podcast)
             }
 
-            if (viewModel.lastPlaybackState != null && (viewModel.lastPlaybackState?.episodeUuid != state.episodeUuid || (viewModel.lastPlaybackState?.isPlaying == false && state.isPlaying)) && settings.openPlayerAutomatically()) {
+            if (viewModel.lastPlaybackState != null &&
+                (
+                    viewModel.lastPlaybackState?.episodeUuid != state.episodeUuid ||
+                        (viewModel.lastPlaybackState?.isPlaying == false && state.isPlaying)
+                    ) &&
+                settings.openPlayerAutomatically.value
+            ) {
                 binding.playerBottomSheet.openPlayer()
             }
 
@@ -800,7 +806,7 @@ class MainActivity :
     private fun updatePlaybackState(state: PlaybackState) {
         binding.playerBottomSheet.setPlaybackState(state)
 
-        if ((state.isPlaying || state.isBuffering) && settings.keepScreenAwake()) {
+        if ((state.isPlaying || state.isBuffering) && settings.keepScreenAwake.value) {
             window?.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
         } else {
             window?.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)