+++ modules/features/account/src/main/java/au/com/shiftyjelly/pocketcasts/account/PromoCodeFragment.kt
@@ -64,8 +64,7 @@ class PromoCodeFragment : BaseFragment() {
                         binding.loadedGroup.isVisible = false
                         // We need to set this to true here so we know not to show the gift dialog.
                         // There is no way to know what type of gift the upgrade came via the API
-                        settings.setFreeGiftAcknowledged(true)
-                        settings.setFreeGiftAcknowledgedNeedsSync(true)
+                        settings.freeGiftAcknowledged.set(true, needsSync = true)
 
                         val result = Intent()
                         result.putExtra(AccountActivity.PROMO_CODE_RETURN_DESCRIPTION, it.response.description)