+++ modules/features/account/src/main/java/au/com/shiftyjelly/pocketcasts/account/viewmodel/OnboardingWelcomeViewModel.kt
@@ -34,8 +34,7 @@ class OnboardingWelcomeViewModel @Inject constructor(
             )
         )
 
-        settings.setMarketingOptIn(newsletter)
-        settings.setMarketingOptInNeedsSync(true)
+        settings.marketingOptIn.set(newsletter, needsSync = true)
     }
 
     fun onShown(flow: OnboardingFlow) {