+++ modules/features/profile/src/main/java/au/com/shiftyjelly/pocketcasts/profile/cloud/CloudFileBottomSheet.kt
@@ -336,7 +336,7 @@ class CloudFileBottomSheetFragment : BottomSheetDialogFragment() {
 
     fun download(episode: UserEpisode, isOnWifi: Boolean) {
         viewModel.trackOptionTapped(DOWNLOAD)
-        if (settings.warnOnMeteredNetwork() && !isOnWifi) {
+        if (settings.warnOnMeteredNetwork.value && !isOnWifi) {
             warningsHelper.downloadWarning(episodeUUID, "user episode sheet")
                 .show(parentFragmentManager, "download_warning")
         } else {
@@ -346,7 +346,7 @@ class CloudFileBottomSheetFragment : BottomSheetDialogFragment() {
 
     private fun upload(episode: UserEpisode, isOnWifi: Boolean) {
         viewModel.trackOptionTapped(UPLOAD)
-        if (settings.warnOnMeteredNetwork() && !isOnWifi) {
+        if (settings.warnOnMeteredNetwork.value && !isOnWifi) {
             warningsHelper.uploadWarning(episodeUUID, source = SourceView.FILES)
                 .show(parentFragmentManager, "upload_warning")
         } else {