+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/AppearanceSettingsFragment.kt
@@ -160,7 +160,7 @@ class AppearanceSettingsFragment : BaseFragment() {
             viewModel.useAndroidLightDarkMode(isChecked, activity as? AppCompatActivity)
         }
 
-        binding.swtShowArtwork.isChecked = viewModel.showArtworkOnLockScreen.value ?: false
+        binding.swtShowArtwork.isChecked = viewModel.showArtworkOnLockScreen.value
         binding.swtShowArtwork.setOnCheckedChangeListener { _, isChecked ->
             viewModel.updateShowArtworkOnLockScreen(isChecked)
         }
@@ -168,7 +168,7 @@ class AppearanceSettingsFragment : BaseFragment() {
             binding.swtShowArtwork.isChecked = !binding.swtShowArtwork.isChecked
         }
 
-        binding.swtUseEmbeddedArtwork.isChecked = viewModel.useEmbeddedArtwork.value ?: false
+        binding.swtUseEmbeddedArtwork.isChecked = viewModel.useEmbeddedArtwork.value
         binding.swtUseEmbeddedArtwork.setOnCheckedChangeListener { _, isChecked ->
             viewModel.updateUseEmbeddedArtwork(isChecked)
         }