+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/AutoArchiveFragment.kt
@@ -1,10 +1,10 @@
 package au.com.shiftyjelly.pocketcasts.settings
 
-import android.content.SharedPreferences
 import android.os.Bundle
 import android.view.View
 import androidx.appcompat.widget.Toolbar
 import androidx.fragment.app.viewModels
+import androidx.preference.ListPreference
 import androidx.preference.PreferenceFragmentCompat
 import androidx.preference.SwitchPreference
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
@@ -14,22 +14,48 @@ import au.com.shiftyjelly.pocketcasts.views.extensions.setup
 import au.com.shiftyjelly.pocketcasts.views.helper.HasBackstack
 import au.com.shiftyjelly.pocketcasts.views.helper.NavigationIcon.BackArrow
 import dagger.hilt.android.AndroidEntryPoint
-import timber.log.Timber
 import javax.inject.Inject
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 @AndroidEntryPoint
-class AutoArchiveFragment : PreferenceFragmentCompat(), HasBackstack, SharedPreferences.OnSharedPreferenceChangeListener {
+class AutoArchiveFragment : PreferenceFragmentCompat(), HasBackstack {
     @Inject lateinit var settings: Settings
     @Inject lateinit var theme: Theme
 
     private val viewModel: AutoArchiveFragmentViewModel by viewModels()
+
+    private lateinit var autoArchivePlayedEpisodes: ListPreference
+    private lateinit var autoArchiveInactiveEpisodes: ListPreference
+    private lateinit var autoArchiveIncludeStarred: SwitchPreference
 
     val toolbar: Toolbar?
         get() = view?.findViewById(R.id.toolbar)
 
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         addPreferencesFromResource(R.xml.preferences_auto_archive)
+        autoArchivePlayedEpisodes = preferenceManager.findPreference<ListPreference>("autoArchivePlayedEpisodes")!!
+            .apply {
+                setOnPreferenceChangeListener { _, newValue ->
+                    viewModel.onPlayedEpisodesAfterChanged(newValue as String)
+                    true
+                }
+            }
+        autoArchiveInactiveEpisodes = preferenceManager.findPreference<ListPreference>("autoArchiveInactiveEpisodes")!!
+            .apply {
+                setOnPreferenceChangeListener { _, newValue ->
+                    viewModel.onInactiveChanged(newValue as String)
+                    true
+                }
+            }
+
+        autoArchiveIncludeStarred = preferenceManager.findPreference<SwitchPreference>("autoArchiveIncludeStarred")!!
+            .apply {
+                setOnPreferenceChangeListener { _, newValue ->
+                    viewModel.onStarredChanged(newValue as Boolean)
+                    updateStarredSummary()
+                    true
+                }
+            }
         updateStarredSummary()
     }
 
@@ -41,35 +67,19 @@ class AutoArchiveFragment : PreferenceFragmentCompat(), HasBackstack, SharedPref
 
     override fun onResume() {
         super.onResume()
-        preferenceScreen.sharedPreferences?.registerOnSharedPreferenceChangeListener(this)
+        setupAutoArchiveAfterPlaying()
+        setupAutoArchiveInactive()
+        setupIncludeStarred()
     }
 
     override fun onPause() {
         super.onPause()
-        preferenceScreen.sharedPreferences?.unregisterOnSharedPreferenceChangeListener(this)
         viewModel.onFragmentPause(activity?.isChangingConfigurations)
     }
 
     private fun updateStarredSummary() {
-        val starredSummary = getString(if (settings.getAutoArchiveIncludeStarred()) LR.string.settings_auto_archive_starred_summary else LR.string.settings_auto_archive_no_starred_summary)
-        val preference = preferenceManager.findPreference<SwitchPreference>(Settings.AUTO_ARCHIVE_INCLUDE_STARRED)
-        preference?.summary = starredSummary
-    }
-
-    override fun onSharedPreferenceChanged(sharedPreferences: SharedPreferences?, key: String?) {
-        when (key) {
-            Settings.AUTO_ARCHIVE_INCLUDE_STARRED -> {
-                updateStarredSummary()
-                viewModel.onStarredChanged()
-            }
-            Settings.AUTO_ARCHIVE_PLAYED_EPISODES_AFTER -> {
-                viewModel.onPlayedEpisodesAfterChanged()
-            }
-            Settings.AUTO_ARCHIVE_INACTIVE -> {
-                viewModel.onInactiveChanged()
-            }
-            else -> Timber.d("Unknown preference changed: $key")
-        }
+        val starredSummary = getString(if (settings.autoArchiveIncludeStarred.value) LR.string.settings_auto_archive_starred_summary else LR.string.settings_auto_archive_no_starred_summary)
+        autoArchiveIncludeStarred.summary = starredSummary
     }
 
     override fun onBackPressed(): Boolean {
@@ -84,5 +94,19 @@ class AutoArchiveFragment : PreferenceFragmentCompat(), HasBackstack, SharedPref
 
     override fun getBackstackCount(): Int {
         return childFragmentManager.backStackEntryCount
+    }
+
+    private fun setupAutoArchiveAfterPlaying() {
+        val stringArray = resources.getStringArray(LR.array.settings_auto_archive_played_values)
+        autoArchivePlayedEpisodes.value = stringArray[settings.autoArchiveAfterPlaying.value.toIndex()]
+    }
+
+    private fun setupAutoArchiveInactive() {
+        val stringArray = resources.getStringArray(LR.array.settings_auto_archive_inactive_values)
+        autoArchiveInactiveEpisodes.value = stringArray[settings.autoArchiveInactive.value.toIndex()]
+    }
+
+    private fun setupIncludeStarred() {
+        autoArchiveIncludeStarred.isChecked = settings.autoArchiveIncludeStarred.value
     }
 }