+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/privacy/PrivacyViewModel.kt
@@ -33,8 +33,8 @@ class PrivacyViewModel @Inject constructor(
     private val mutableUiState = MutableStateFlow<UiState>(
         UiState.Loaded(
             analytics = analyticsTracker.getSendUsageStats(),
-            crashReports = settings.getSendCrashReports(),
-            linkAccount = settings.getLinkCrashReportsToUser(),
+            crashReports = settings.sendCrashReports.value,
+            linkAccount = settings.linkCrashReportsToUser.value,
             getUserEmail = { syncManager.getEmail() }
         )
     )