+++ wear/src/main/kotlin/au/com/shiftyjelly/pocketcasts/wear/ui/settings/SettingsViewModel.kt
@@ -37,8 +37,8 @@ class SettingsViewModel @Inject constructor(
         State(
             refreshState = null,
             signInState = userManager.getSignInState().blockingFirst(),
-            showDataWarning = settings.warnOnMeteredNetwork(),
-            refreshInBackground = settings.refreshPodcastsAutomatically(),
+            showDataWarning = settings.warnOnMeteredNetwork.value,
+            refreshInBackground = settings.backgroundRefreshPodcasts.value,
         )
     )
     val state = _state.asStateFlow()
@@ -62,12 +62,12 @@ class SettingsViewModel @Inject constructor(
     }
 
     fun setWarnOnMeteredNetwork(warnOnMeteredNetwork: Boolean) {
-        settings.setWarnOnMeteredNetwork(warnOnMeteredNetwork)
+        settings.warnOnMeteredNetwork.set(warnOnMeteredNetwork)
         _state.update { it.copy(showDataWarning = warnOnMeteredNetwork) }
     }
 
     fun setRefreshPodcastsInBackground(isChecked: Boolean) {
-        settings.setRefreshPodcastsAutomatically(isChecked)
+        settings.backgroundRefreshPodcasts.set(isChecked)
         _state.update { it.copy(refreshInBackground = isChecked) }
     }